rename main to app remove unnecessary method in meta service
